Move rig-macros contents into rig-core-derive
