+++ rig-core/rig-core-derive/Cargo.toml
@@ -7,6 +7,7 @@ description = "Internal crate that implements Rig derive macros."
 repository = "https://github.com/0xPlaygrounds/rig"
 
 [dependencies]
+convert_case = { version = "0.6.0" }
 indoc = "2.0.5"
 proc-macro2 = { version = "1.0.87", features = ["proc-macro"] }
 quote = "1.0.37"
@@ -14,3 +15,25 @@ syn = { version = "2.0.79", features = ["full"]}
 
 [lib]
 proc-macro = true
+
+[dev-dependencies]
+rig-core = { path = "../../rig-core" }
+serde = "1.0"
+tokio = { version = "1.44.0", features = ["full"] }
+tracing-subscriber = "0.3.0"
+
+[[example]]
+name = "simple"
+path = "examples/rig_tool/simple.rs"
+
+[[example]]
+name = "with_description"
+path = "examples/rig_tool/with_description.rs"
+
+[[example]]
+name = "full"
+path = "examples/rig_tool/full.rs"
+
+[[example]]
+name = "async_tool"
+path = "examples/rig_tool/async_tool.rs"
