--- rig-macros/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "rig-macros"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "2.0.100" }
-quote = { version = "1.0.40" }
-convert_case = { version = "0.6.0" }
-rig-core = { path = "../rig-core" }
-proc-macro2 = "1.0.94"
-
-[dev-dependencies]
-serde = "1.0"
-tokio = { version = "1.44.0", features = ["full"] }
-tracing-subscriber = "0.3.0"
-
-[[example]]
-name = "simple"
-path = "examples/simple.rs"
-
-[[example]]
-name = "with_description"
-path = "examples/with_description.rs"
-
-[[example]]
-name = "full"
-path = "examples/full.rs"
-
-[[example]]
-name = "async_tool"
-path = "examples/async_tool.rs"
